Добавлен интерфейс IUserRepository и класс UserRepository, реализующий связь с таблицей users. Создан отдельный контроллер для действий с пользователями.
